Moves decorated_types.rs into the api layer
