+++ full-service/src/api/wallet_api.rs
@@ -319,8 +319,8 @@ pub fn rocket(rocket_config: rocket::Config, state: WalletApiState) -> rocket::R
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::api::decorated_types::{JsonAccount, JsonCreateAccountResponse};
     use crate::service::MockWallet;
-    use crate::service::{JsonAccount, JsonCreateAccountResponse};
     use mc_common::logger::{test_with_logger, Logger};
     use rocket::{
         http::{ContentType, Status},
