+++ full-service/src/db/account.rs
@@ -17,8 +17,8 @@ use mc_account_keys::{AccountKey, DEFAULT_SUBADDRESS_INDEX};
 use mc_crypto_digestible::{Digestible, MerlinTranscript};
 use mc_transaction_core::ring_signature::KeyImage;
 
+use crate::api::decorated_types::JsonAccount;
 use crate::db::WalletDbError;
-use crate::service::JsonAccount;
 use diesel::{
     prelude::*,
     r2d2::{ConnectionManager, PooledConnection},
